+++ plugins/touch-music/src/components/music/word-lyric/WordLyricScroller.vue
@@ -65,7 +65,7 @@ async function handleIndex(i) {
 
   clearTimeout(task)
   task = setTimeout(async () => {
-    const el = scroll.value.$el.children[0].children[0]
+    const el = scroll.value?.$el?.querySelector('.tx-scroll__content')
     if (!el)
       return
     const target = el.children[i].offsetTop - 100
@@ -78,19 +78,19 @@ async function handleIndex(i) {
 
 <template>
   <div class="WordLyric-Container" :class="{ shine }" :style="`--theme-word-color: ${song.colors && song.colors[song.colors.length > 4 ? 1 : 0].color}`">
-    <el-scrollbar ref="scroll">
+    <TxScroll ref="scroll">
       <WordLyricItem
         v-for="(item, index) in wordLyrics || []" :key="index"
         :tlyric="tlyric[index]"
         :class="{ 'start': index === ind, 'far-away': index + 2 === ind || index - 2 === ind, 'far': index + 1 === ind || index - 1 === ind }" :index="index" :lyric="item" @index="handleIndex"
       />
-    </el-scrollbar>
+    </TxScroll>
   </div>
 </template>
 
 <style lang="scss" scoped>
 .WordLyric-Container {
-  :deep(.el-scrollbar__view) {
+  :deep(.tx-scroll__content) {
     width: 300px;
   }
   position: relative;